+++ src/main/java/pl/coderslab/charity/repository/CategoryRepository.java
@@ -3,5 +3,5 @@ package pl.coderslab.charity.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.coderslab.charity.entity.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Double> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 }
