+++ src/main/java/pl/coderslab/charity/repository/DonationRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import pl.coderslab.charity.entity.Donation;
 
-public interface DonationRepository extends JpaRepository<Donation, Double> {
+public interface DonationRepository extends JpaRepository<Donation, Long> {
     @Query("SELECT SUM(donation.quantity) FROM Donation donation")
     Integer sumOfBugs();
 }
