+++ src/main/java/pl/coderslab/charity/repository/InstitutionRepository.java
@@ -3,5 +3,5 @@ package pl.coderslab.charity.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.coderslab.charity.entity.Institution;
 
-public interface InstitutionRepository extends JpaRepository<Institution, Double> {
+public interface InstitutionRepository extends JpaRepository<Institution, Long> {
 }
